com.unity.sentis@2.1.1 ## [2.1.1] - 2024-11-07
### Fixed
- Ordering of inputs now matches ONNX model consistently
- Loading of models with constants over 500mb
- Serialization of models over 2GB in size
- Issues in example code in documentation
- Shape inference for `Split` and `Select` operators
- Profiler tags for operators are now consistent with lower overhead
- `CopyOutput` can now be used on GPUPixel backend if the tensor is the correct shape

+++ Editor/ONNX/ONNXConstantsLoader.cs
@@ -149,7 +149,7 @@ namespace Unity.Sentis.ONNX
             else if (dataType == TensorProto.Types.DataType.Float)
             {
                 Assert.IsTrue((sizeof(float) * shape.length) == length);
-                data = new NativeTensorArrayFromManagedArray(byteArray, 0, sizeof(float), shape.length);
+                data = new NativeTensorArrayFromManagedArray(byteArray, 0, sizeof(byte), shape.length);
             }
             // Float16
             else if (dataType == TensorProto.Types.DataType.Float16)
@@ -175,7 +175,7 @@ namespace Unity.Sentis.ONNX
             else if (dataType == TensorProto.Types.DataType.Int32)
             {
                 Assert.IsTrue((sizeof(int) * shape.length) == length);
-                data = new NativeTensorArrayFromManagedArray(byteArray, 0, sizeof(int), shape.length);
+                data = new NativeTensorArrayFromManagedArray(byteArray, 0, sizeof(byte), shape.length);
             }
             // Int64
             else if (dataType == TensorProto.Types.DataType.Int64)
